// server/src/permissions/permission.enum.ts
/* eslint-disable prettier/prettier */
// src/common/enums/permissions.enum.ts
export enum Permission {
  CREATE_RESTAURANT = 'create_restaurant',
  UPDATE_RESTAURANT = 'update_restaurant',
  DELETE_RESTAURANT = 'delete_restaurant',
  VIEW_RESTAURANT = 'view_restaurant',

  CREATE_TABLE = 'create_table',
  UPDATE_TABLE = 'update_table',
  DELETE_TABLE = 'delete_table',

  ADD_SEAT = 'add_seat',
  DELETE_SEAT = 'delete_seat',

  CREATE_MENU = 'create_menu',
  UPDATE_MENU = 'update_menu',
  DELETE_MENU = 'delete_menu',
  VIEW_ORDERS = 'view_orders'
}